Restructure PR review workspace layout
- Add dedicated panes for PR list, file tabs, inspector, and conflict review
- Introduce comment composer, thread cards, and mention-aware comment rendering
- Improve workspace navigation and focus controls for review workflows
